app/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.26.164
botocore==1.29.165
dj-database-url==2.0.0
Django==4.2.2
django-cors-headers==4.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
